+++ pkg/bootstrap/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["authenticate.go"],
+    importpath = "k8s.io/kops/pkg/bootstrap",
+    visibility = ["//visibility:public"],
+)
