+++ upup/pkg/fi/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "assetstore.go",
-        "authenticate.go",
         "ca.go",
         "changes.go",
         "clientset_castore.go",
