+++ upup/pkg/fi/cloudup/awsup/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "//dnsprovider/pkg/dnsprovider/providers/aws/route53:go_default_library",
         "//pkg/apis/kops:go_default_library",
         "//pkg/apis/kops/model:go_default_library",
+        "//pkg/bootstrap:go_default_library",
         "//pkg/cloudinstances:go_default_library",
         "//pkg/featureflag:go_default_library",
         "//pkg/nodeidentity/aws:go_default_library",
