+++ upup/pkg/fi/cloudup/awsup/aws_verifier.go
@@ -36,8 +36,8 @@ import (
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/sts"
+	"k8s.io/kops/pkg/bootstrap"
 	nodeidentityaws "k8s.io/kops/pkg/nodeidentity/aws"
-	"k8s.io/kops/upup/pkg/fi"
 )
 
 type AWSVerifierOptions struct {
@@ -57,9 +57,9 @@ type awsVerifier struct {
 	client http.Client
 }
 
-var _ fi.Verifier = &awsVerifier{}
+var _ bootstrap.Verifier = &awsVerifier{}
 
-func NewAWSVerifier(opt *AWSVerifierOptions) (fi.Verifier, error) {
+func NewAWSVerifier(opt *AWSVerifierOptions) (bootstrap.Verifier, error) {
 	config := aws.NewConfig().
 		WithCredentialsChainVerboseErrors(true).
 		WithRegion(opt.Region).
@@ -120,7 +120,7 @@ type ResponseMetadata struct {
 	RequestId string `xml:"RequestId"`
 }
 
-func (a awsVerifier) VerifyToken(token string, body []byte) (*fi.VerifyResult, error) {
+func (a awsVerifier) VerifyToken(token string, body []byte) (*bootstrap.VerifyResult, error) {
 	if !strings.HasPrefix(token, AWSAuthenticationTokenPrefix) {
 		return nil, fmt.Errorf("incorrect authorization type")
 	}
@@ -237,7 +237,7 @@ func (a awsVerifier) VerifyToken(token string, body []byte) (*fi.VerifyResult, e
 		return nil, err
 	}
 
-	result := &fi.VerifyResult{
+	result := &bootstrap.VerifyResult{
 		NodeName:         addrs[0],
 		CertificateNames: addrs,
 	}
