+++ upup/pkg/fi/nodeup/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//pkg/apis/kops/registry:go_default_library",
         "//pkg/apis/nodeup:go_default_library",
         "//pkg/assets:go_default_library",
+        "//pkg/bootstrap:go_default_library",
         "//pkg/configserver:go_default_library",
         "//pkg/kopscodecs:go_default_library",
         "//upup/pkg/fi:go_default_library",
