+++ upup/pkg/fi/nodeup/nodetasks/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
         "//pkg/apis/kops:go_default_library",
         "//pkg/apis/nodeup:go_default_library",
         "//pkg/backoff:go_default_library",
+        "//pkg/bootstrap:go_default_library",
         "//pkg/kubeconfig:go_default_library",
         "//pkg/pki:go_default_library",
         "//upup/pkg/fi:go_default_library",
